Added delete methods & more loggings informations

+++ src/api/playlists.ts
@@ -1,9 +1,9 @@
 import { Client } from '..';
 
-import { AddedSong, OnDupes, OrderDirections, OrderTypes, TidalPlaylist, TidalSong } from '../types';
+import { AddedSong, Deleted, OnDupes, OrderDirections, OrderTypes, TidalPlaylist, TidalSong } from '../types';
 
 export class Playlists {
-  client: Client;
+  private client: Client;
 
   constructor(client: Client) {
     this.client = client;
@@ -27,7 +27,7 @@ export class Playlists {
    * @param {number} [offset=0] - The offset of the first song to return.
    * @param {OrderTypes} [order=DATE] - OrderTypes = 'DATE',
    * @param {OrderDirections} [orderDirection=ASC] - OrderDirections = 'ASC',
-   * @returns An array of songs
+   * @returns TidalSong
    */
   public async getPlaylistSongs(
     playlistId: string,
@@ -46,6 +46,48 @@ export class Playlists {
       },
     });
     return response as TidalSong;
+  }
+
+  /**
+   * It deletes a playlist from the user's account.
+   * @param {string} playlistId - The id of the playlist you want to delete
+   * @returns Deleted.
+   */
+  public async deletePlaylist(playlistId: string) {
+    if (!playlistId) throw new Error('PlaylistId not specified');
+    const playlist = await this.getPlaylistInfos(playlistId);
+    if (!playlist) throw new Error(`There is no playlist with the ${playlist.uuid} id`);
+    const response = await this.client.request(`my-collection/playlists/folders/remove`, {
+      modes: 'api',
+      method: 'PUT',
+      versions: 'v2',
+      params: {
+        trns: `trn:playlist:${playlist.uuid}`,
+      },
+    });
+    if (!response) return { status: 'Success', playlistId, playlist } as Deleted;
+    else return { status: 'Failed', playlistId, playlist } as Deleted;
+  }
+
+  /**
+   * It takes a folderId as a parameter, and returns a Deleted object with a status of 'Success' or
+   * 'Failed' and the folderId.
+   * @param {string} folderId - The id of the folder you want to delete
+   * @returns Deleted.
+   */
+  public async deleteFolder(folderId: string) {
+    if (!folderId) throw new Error('PlaylistId not specified');
+    const response = await this.client.request(`my-collection/playlists/folders/remove`, {
+      modes: 'api',
+      method: 'PUT',
+      versions: 'v2',
+      params: {
+        trns: `trn:folder:${folderId}`,
+      },
+    });
+    console.log(response);
+    if (!response) return { status: 'Success', folderId } as Deleted;
+    else return { status: 'Failed', folderId } as Deleted;
   }
 
   /**
@@ -74,12 +116,12 @@ export class Playlists {
   }
 
   /**
-   * It deletes a song from a playlist.
-   * @param {string} playlistId - The ID of the playlist you want to delete a song from.
-   * @param {number} index - The index of the song in the playlist
+   * It deletes a song from a playlist
+   * @param {string} playlistId - The id of the playlist you want to delete a song from
+   * @param {number} index - The index of the song you want to delete
    * @param {OrderTypes} [order=INDEX] - OrderTypes = 'INDEX'
    * @param {OrderDirections} [orderDirection=ASC] - OrderDirections = 'ASC'
-   * @returns The response from the server.
+   * @returns Deleted.
    */
   public async deleteSong(
     playlistId: string,
@@ -89,6 +131,9 @@ export class Playlists {
   ) {
     if (!playlistId) throw new Error('PlaylistId not specified');
     if (!index) throw new Error('Index not specified');
+    const { numberOfTracks } = await this.getPlaylistInfos(playlistId);
+    if (numberOfTracks <= 1) throw new Error('The playlist need to contain at least 2 tracks');
+    const { items } = await this.getPlaylistSongs(playlistId);
     const response = await this.client.request(`playlists/${playlistId}/items/${index}`, {
       method: 'DELETE',
       headers: {
@@ -99,6 +144,7 @@ export class Playlists {
         orderDirection,
       },
     });
-    return response;
+    if (!response) return { status: 'Success', playlistId, index, song: items[index].item } as Deleted;
+    else return { status: 'Failed', playlistId, index } as Deleted;
   }
 }
